% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addError.R
\name{addError}
\alias{addError}
\title{Generates error using SD derived from lowess model}
\usage{
addError(xpred, lowess_obj)
}
\arguments{
\item{xpred}{a predicted x value}

\item{lowess_obj}{a list containing output from the lowess function}
}
\value{
A random draw from a normal distribution with mean zero and sd derived from lowess object
}
\description{
Description text
}
