% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAR.R
\name{buildAR}
\alias{buildAR}
\title{Builds an autoregressive model}
\usage{
buildAR(vec, wsize, method = c("equal", "triangle"))
}
\arguments{
\item{vec}{A vector of numeric data}

\item{wsize}{Number of prior observations to use for averaging}

\item{method}{Type of weighting to use}
}
\value{
A list with the following components:

wsize is the number of prior observations used for averaging

vec is a numeric vector containing the input data

fit is a numeric vector containing the autoregressive model fitted data

errors is a numeric vector containing the difference between fit and vec

initphi is a numeric vector containing the initial phi

finalphi is a numeric vector containing the final phi
}
\description{
Description text
}
