% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiedPegelsfcastC.R
\name{modifiedPegelsfcastC}
\alias{modifiedPegelsfcastC}
\title{Modified version of pegelsfcast.C}
\usage{
modifiedPegelsfcastC(h, obj, npaths, level, bootstrap)
}
\arguments{
\item{h}{How far in the future to forecast.}

\item{obj}{A mode object hard-coded to the Holt model.}

\item{npaths}{Number of simulations.}

\item{level}{Quantile of distribution. Not utilized but maintained for consistency.}

\item{bootstrap}{Whether to resample errors (TRUE) or use Gaussian errors (FALSE).}
}
\value{
A list with the following components:

mu is the mean path

lower is the lower quantile

upper is the upper quantile

y_paths is a matrix of paths
}
\description{
This function, originally called pegelsfcast.C from the forecast package,
implements the work in Forecasting: Principles and Practice by
Rob J Hyndman and George Athanasopoulos (https://otexts.com/fpp2/)
}
