% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictAR.R
\name{predictAR}
\alias{predictAR}
\title{Makes predictions using an autogregressive model from the buildAR function}
\usage{
predictAR(
  buildAR_obj,
  pdays = 28,
  nsim = 100,
  skip = 0,
  seed = NULL,
  output_type = "all",
  debug = FALSE
)
}
\arguments{
\item{buildAR_obj}{An object output from the buildAR function}

\item{pdays}{Number of days into the future to make predictions, default is 28}

\item{nsim}{Number of simulations, default is 100}

\item{skip}{Number of input values to skip, default is 0}

\item{seed}{Seed for random number generator}

\item{output_type}{Type of output to predict, currently max, mean, or all are valid}
}
\value{
A data frame containing the specified output for each sim
}
\description{
Description text
}
