% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressARLag.R
\name{regressARLag}
\alias{regressARLag}
\title{Combines multiple autoregressive models into one prediction model using a least sqares regression model}
\usage{
regressARLag(
  vec,
  x = NULL,
  x_lag = NULL,
  index_colname = "date",
  output_type = "Max",
  wsize,
  method = c("unweighted", "equal", "triangle"),
  pdays,
  nsim,
  skip = 0,
  seed = NULL,
  regression_weights = NULL,
  rhat_method = c("none", "geometric", "arithmetic"),
  debug = FALSE
)
}
\arguments{
\item{vec}{A vector of numeric data, will also be include as a default predictor}

\item{x}{A data frame containing covariates with which to generate predictive model, if unspecified, defaults to vec}

\item{x_lag}{An array with an element for each non-date column in x, indicating the lag associated with that column}

\item{index_colname}{An character indicating the name of the column that temporally indexes data, e.g. "date". This column should be yyyy/mm/dd format}

\item{output_type}{A string indicating which outcome measure should be predicted. Must be one of Min, FirstQu, Median, Mean, ThirdQu, or Max}

\item{wsize}{Number of prior observations to use for averaging}

\item{method}{Type of weighting to use in individual prediction models}

\item{pdays}{Number of days into the future to make predictions}

\item{nsim}{Number of simulations}

\item{skip}{Number of input values to skip}

\item{seed}{Seed for random number generator}

\item{regression_weights}{An array specifying regression weights. Currently not implemented}

\item{rhat_method}{Method for calculating rhat, if "none", rhat = 1 and has no effect}

\item{debug}{TRUE returns buildAR objects in addition to standard output}
}
\value{
A list containing the specified output statistics for each sim
}
\description{
Need to fill in description data
}
