% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateHoltMA.R
\name{simulateHoltMA}
\alias{simulateHoltMA}
\title{simulateHoltMA}
\usage{
simulateHoltMA(
  dat,
  h,
  npaths,
  level = 0.95,
  bootstrap,
  ma,
  smooth_type = "s",
  model_type = "AAN",
  seed = NULL
)
}
\arguments{
\item{dat}{One-dimensional vector of interest like test positivity}

\item{h}{How far in the future to forecast}

\item{npaths}{Number of simulations}

\item{level}{Prediction level, not used but need for existing code}

\item{bootstrap}{Whether to resample errors (TRUE) or use Gaussian errors (FALSE)}

\item{ma}{Width of moving average window in days}

\item{smooth_type}{Type of smoothing used by movavg, see movavg documentation for detials}

\item{model_type}{From the ets documentation: Usually a three-character string identifying method using the framework terminology of Hyndman et al. (2002) and Hyndman et al. (2008). The first letter denotes the error type ("A", "M" or "Z"); the second letter denotes the trend type ("N","A","M" or "Z"); and the third letter denotes the season type ("N","A","M" or "Z"). In all cases, "N"=none, "A"=additive, "M"=multiplicative and "Z"=automatically selected. So, for example, "ANN" is simple exponential smoothing with additive errors, "MAM" is multiplicative Holt-Winters' method with multiplicative errors, and so on.}

\item{seed}{Random seed to use}
}
\value{
A list with the following components:

paths is raw matrix of paths including training data

paths_ma is moving average matrix of paths

smoothed_training is moving average of training data

max_predictions is maximum for each row of paths_ma
}
\description{
Need a better description of this function
}
