% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAR.R
\name{buildAR}
\alias{buildAR}
\title{Builds an autoregressive model}
\usage{
buildAR(
  vec,
  x = NULL,
  wsize,
  method = c("unweighted", "equal", "triangle"),
  seed = NULL,
  rhat_method = c("none", "geometric", "arithmetic"),
  lambda = 0
)
}
\arguments{
\item{vec}{A vector of numeric data}

\item{x}{A vector containing a covariate with which to generate AR model, if unspecified, defaults to vec}

\item{wsize}{Number of prior observations to use for averaging}

\item{method}{Type of weighting to use}

\item{seed}{Seed for random number generator}

\item{rhat_method}{Method for calculating rhat, if "none", rhat = 1 and has no effect}

\item{lambda}{Shrinkage parameter, if not specified, default is 0 which produces no shrinkage. If an array is specified, all values in teh array
are evaluated and the optimal lambda is chosen based on residual sum of squares. Values should be between 0 and 1 inclusive.}
}
\value{
A list with the following components:

wsize is the number of prior observations used for averaging

vec is a numeric vector containing the input data

fit is a numeric vector containing the autoregressive model fitted data

errors is a numeric vector containing the difference between fit and vec

initphi is a numeric vector containing the initial phi

finalphi is a numeric vector containing the final phi
}
\description{
Description text
}
