% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictAR.R
\name{predictAR}
\alias{predictAR}
\title{Makes predictions using an autogregressive model from the buildAR function}
\usage{
predictAR(buildAR_obj, pdays, nsim, skip = 0, seed = NULL, output_type = "max")
}
\arguments{
\item{buildAR_obj}{An object output from the buildAR function}

\item{pdays}{Number of days into the future to make predictions}

\item{nsim}{Number of simulations}

\item{skip}{Number of input values to skip}

\item{seed}{Seed for random number generator}

\item{output_type}{Type of output to predict, currently max, mean, or all are valid}
}
\value{
A data frame containing the specified output for each sim
}
\description{
Description text
}
