% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressAR.R
\name{regressAR}
\alias{regressAR}
\title{Combines multiple autoregressive models into one prediction model using a least sqares regression model}
\usage{
regressAR(
  vec,
  x = NULL,
  output_type = "max",
  wsize = 14,
  method = c("equal", "unweighted", "triangle"),
  pdays = 28,
  nsim = 100,
  skip = 0,
  seed = NULL,
  regression_weights = NULL,
  rhat_method = c("none", "geometric", "arithmetic"),
  debug = FALSE
)
}
\arguments{
\item{vec}{A vector of numeric data, will also be include as a default predictor}

\item{x}{A data frame containing covariates with which to generate predictive model, if unspecified, defaults to vec}

\item{output_type}{A string indicating which outcome measure should be predicted. Must be one of Min, FirstQu, Median, Mean, ThirdQu, or Max}

\item{wsize}{Number of prior observations to use for averaging, default is 14}

\item{method}{Type of weighting to use in individual prediction models, default is equal}

\item{pdays}{Number of days into the future to make predictions, default is 28}

\item{nsim}{Number of simulations, default is 100}

\item{skip}{Number of input values to skip, default is 0}

\item{seed}{Seed for random number generator}

\item{regression_weights}{An array specifying regression weights. Currently not implemented}

\item{rhat_method}{Method for calculating rhat, if "none", rhat = 1 and has no effect}

\item{debug}{TRUE returns buildAR objects in addition to standard output}
}
\value{
A list containing the specified output statistics for each sim
}
\description{
Need to fill in description data
}
