% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressAR.R
\name{regressAR}
\alias{regressAR}
\title{Combines multiple autoregressive models in a least sqares regression model}
\usage{
regressAR(
  vec,
  x = NULL,
  predict_variable = "Max",
  wsize,
  method = c("unweighted", "equal", "triangle"),
  pdays,
  nsim,
  skip = 0,
  seed = NULL,
  output_type = "all",
  regression_weights = NULL
)
}
\arguments{
\item{vec}{A vector of numeric data}

\item{x}{A data frame containing covariates with which to generate predictive model, if unspecified, defaults to vec}

\item{predict_variable}{A string indicating which outcome measure should be predicted. Must be one of Min, FirstQu, Median, Mean, ThirdQu, or Max}

\item{wsize}{Number of prior observations to use for averaging}

\item{method}{Type of weighting to use}

\item{pdays}{Number of days into the future to make predictions}

\item{nsim}{Number of simulations}

\item{skip}{Number of input values to skip}

\item{seed}{Seed for random number generator}

\item{output_type}{Type of output to predict, currently max, mean, or all are valid}

\item{regression_weights}{An array specifying regression weights. Currently not implemented}
}
\value{
A data frame containing the specified output statistics for each sim
}
\description{
Need to fill in description data
}
