% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateAR.R
\name{simulateAR}
\alias{simulateAR}
\title{Builds an autoregressive model and uses it to predict}
\usage{
simulateAR(
  vec,
  x = NULL,
  wsize,
  method = c("unweighted", "equal", "triangle"),
  pdays,
  nsim,
  skip = 0,
  seed = NULL,
  output_type = "all",
  rhat_method = c("none", "geometric", "arithmetic"),
  lambda = 0,
  debug = FALSE
)
}
\arguments{
\item{vec}{A vector of numeric data}

\item{x}{A vector of numeric data, used to predict vec. If not specified, defaults to vec}

\item{wsize}{Number of prior observations to use for averaging}

\item{method}{Type of weighting to use}

\item{pdays}{Number of days into the future to make predictions}

\item{nsim}{Number of simulations}

\item{skip}{Number of input values to skip}

\item{seed}{Seed for random number generator}

\item{output_type}{Type of output}

\item{lambda}{Shrinkage parameter, if not specified, default is 0 which produces no shrinkage. If an array is specified, all values in teh array
are evaluated and the optimal lambda is chosen based on residual sum of squares. Values should be between 0 and 1 inclusive.}

\item{debug}{TRUE returns buildAR objects in addition to standard output}
}
\value{
A data frame containing the specified output statistics for each sim
}
\description{
A wrapper function for buildAR and predictAR.
}
