% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateHoltMA.R
\name{simulateHoltMA}
\alias{simulateHoltMA}
\title{simulate.holt.ma}
\usage{
simulateHoltMA(dat, h, npaths, level, bootstrap, ma)
}
\arguments{
\item{dat}{One-dimensional vector of interest like test positivity.}

\item{h}{How far in the future to forecast.}

\item{npaths}{Number of simulations.}

\item{level}{Prediction level, not used but need for existing code}

\item{bootstrap}{Whether to resample errors (TRUE) or use Gaussian errors (FALSE)}

\item{ma}{Width of moving average window in days}
}
\value{
A list with the following components:

paths is raw matrix of paths including training data

paths.ma is moving average matrix of paths

smoothed.training is moving average of training data

max.predictions is maximum for each row of paths.ma
}
\description{
simulate.holt.ma
}
